Add ignore_configcheck variable so that MPM module switch works.
With Debian Stretch (and presume Jessie, too) the apache2_module tasks
would fail because "apachectl configtest" would fail between disabling
the old MPM module and enabling the new one.

The order of tasks has been also changed so that first comes the
disabling of unwanted MPM module and then after that comes the
enabling of the new one.

This change makes this role require Ansible version 2.3.